backend/requirements.txt
asgiref==3.11.0
Django==4.2.27
django-cors-headers==4.9.0
djangorestframework==3.16.1
django-filter==25.1
djangorestframework-simplejwt==5.3.1
pillow==11.3.0
python-dotenv==1.0.0
sqlparse==0.5.5
typing_extensions==4.15.0
